+++ src/IconRandomizer.cpp
@@ -1,6 +1,5 @@
 #include <Geode/binding/GameManager.hpp>
 #include <Geode/binding/GameStatsManager.hpp>
-#include <Geode/loader/Mod.hpp>
 #include <Geode/utils/ranges.hpp>
 #include <hiimjustin000.more_icons/include/MoreIcons.hpp>
 #include <IconRandomizer.hpp>
@@ -25,6 +24,7 @@ using namespace geode::prelude;
 #define SDI_DEATH "death"
 #define SDI_TRAIL "trail"
 #define SDI_SHIP_FIRE "shiptrail"
+#define SDI_EXPLODE "deathexplode"
 
 // A list of the IDs the player has unlocked for each type of icon.
 std::map<UnlockType, std::vector<int>> UNLOCKED = {
@@ -52,7 +52,7 @@ bool INITIALIZED = false;
 // This takes a parameter for the type of icon to initialize.
 void setupUnlockedIcons(IconType type) {
     // Get the game manager and the vector of unlocked icons for the icon type, then clear the vector.
-    auto gameManager = GameManager::sharedState();
+    auto gameManager = GameManager::get();
     auto& vec = UNLOCKED[gameManager->iconTypeToUnlockType(type)];
     vec.clear();
 
@@ -67,7 +67,7 @@ void setupUnlockedIcons(IconType type) {
 // This takes a parameter for the type of color to initialize.
 void setupUnlockedColors(UnlockType type) {
     // Get the game manager and the vector of unlocked colors for the unlock type, then clear the vector.
-    auto gameManager = GameManager::sharedState();
+    auto gameManager = GameManager::get();
     auto& vec = UNLOCKED[type];
     vec.clear();
 
@@ -75,22 +75,6 @@ void setupUnlockedColors(UnlockType type) {
     for (int i = 0; i < 107; i++) {
         if (gameManager->isColorUnlocked(i, type)) vec.push_back(i);
     }
-}
-
-// A helper function to get a saved value from a Geode mod.
-// This takes the mod, a string for the name of the setting, and a parameter for the default value of the setting.
-template <class T>
-static T getModValue(geode::Mod* mod, std::string const& key, T const& defaultValue) {
-    // Get the mod's saved value if it exists, otherwise return the default value.
-    return mod ? mod->getSavedValue(key, defaultValue) : defaultValue;
-}
-
-// A helper function to set a saved value from a Geode mod.
-// This takes the mod, a string for the name of the setting, and a parameter for the value of the setting.
-template <class T>
-static void setModValue(geode::Mod* mod, std::string const& key, T const& value) {
-    // Set the mod's saved value, if the mod is actually loaded.
-    if (mod) mod->setSavedValue(key, value);
 }
 
 void IconRandomizer::init() {
@@ -166,6 +150,10 @@ RandomizeType IconRandomizer::fromUnlockType(UnlockType type) {
 IconType IconRandomizer::toIconType(RandomizeType type) {
     // Convert the randomize type to an icon type.
     switch (type) {
+        case ICON_RANDOMIZER_API_COLOR_1: return IconType::Item;
+        case ICON_RANDOMIZER_API_COLOR_2: return IconType::Item;
+        case ICON_RANDOMIZER_API_GLOW_COLOR: return IconType::Item;
+        case ICON_RANDOMIZER_API_GLOW: return IconType::Item;
         case ICON_RANDOMIZER_API_CUBE: return IconType::Cube;
         case ICON_RANDOMIZER_API_SHIP: return IconType::Ship;
         case ICON_RANDOMIZER_API_BALL: return IconType::Ball;
@@ -179,6 +167,7 @@ IconType IconRandomizer::toIconType(RandomizeType type) {
         case ICON_RANDOMIZER_API_TRAIL: return IconType::Special;
         case ICON_RANDOMIZER_API_ANIMATION: return IconType::Item;
         case ICON_RANDOMIZER_API_SHIP_FIRE: return IconType::ShipFire;
+        case ICON_RANDOMIZER_API_EXPLODE: return IconType::DeathEffect;
         default: return IconType::Cube;
     }
 }
@@ -203,6 +192,7 @@ UnlockType IconRandomizer::toUnlockType(RandomizeType type) {
         case ICON_RANDOMIZER_API_TRAIL: return UnlockType::Streak;
         case ICON_RANDOMIZER_API_SHIP_FIRE: return UnlockType::ShipFire;
         case ICON_RANDOMIZER_API_ANIMATION: return UnlockType::GJItem;
+        case ICON_RANDOMIZER_API_EXPLODE: return UnlockType::Death;
         default: return UnlockType::Cube;
     }
 }
@@ -222,7 +212,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
 
     // If it is an animation, randomly enable or disable the animations, and return 0.
     if (type == ICON_RANDOMIZER_API_ANIMATION) {
-        auto gameStatsManager = GameStatsManager::sharedState();
+        auto gameStatsManager = GameStatsManager::get();
         for (int i = 0; i < vec.size(); i++) {
             gameStatsManager->toggleEnableItem(UnlockType::GJItem, vec[i], random(0, 1));
         }
@@ -231,37 +221,38 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
 
     // Get the game manager and the Separate Dual Icons mod. If the mod is enabled and the dual parameter is true, use separate dual icons.
     // Then get a random number from the vector of unlocked icons.
-    auto gameManager = GameManager::sharedState();
+    auto gameManager = GameManager::get();
     auto separateDualIcons = Loader::get()->getLoadedMod(SDI_ID);
     auto useDual = separateDualIcons && dual;
     auto loadedIcons = MoreIcons::vectorForType(toIconType(type));
     auto idx = random(0, vec.size() + loadedIcons.size() - 1);
     auto num = idx < vec.size() ? vec[idx] : idx - vec.size();
     auto glow = type == ICON_RANDOMIZER_API_GLOW && ranges::contains(UNLOCKED[UnlockType::Streak], 2) ? (bool)random(0, 1) : false;
+    auto explode = type == ICON_RANDOMIZER_API_EXPLODE ? (bool)random(0, 1) : false;
 
     // Then, set the icon to the random number.
     // If Separate Dual Icons is enabled and the dual parameter is true, set the value in the mod.
     // Otherwise, set the value in the game manager.
     switch (type) {
         case ICON_RANDOMIZER_API_COLOR_1:
-            if (useDual) setModValue(separateDualIcons, SDI_COLOR_1, num);
+            if (useDual) separateDualIcons->setSavedValue(SDI_COLOR_1, num);
             else gameManager->setPlayerColor(num);
             return num;
         case ICON_RANDOMIZER_API_COLOR_2:
-            if (useDual) setModValue(separateDualIcons, SDI_COLOR_2, num);
+            if (useDual) separateDualIcons->setSavedValue(SDI_COLOR_2, num);
             else gameManager->setPlayerColor2(num);
             return num;
         case ICON_RANDOMIZER_API_GLOW_COLOR:
-            if (useDual) setModValue(separateDualIcons, SDI_GLOW_COLOR, num);
+            if (useDual) separateDualIcons->setSavedValue(SDI_GLOW_COLOR, num);
             else gameManager->setPlayerColor3(num);
             return num;
         case ICON_RANDOMIZER_API_GLOW:
-            if (useDual) setModValue(separateDualIcons, SDI_GLOW, glow);
+            if (useDual) separateDualIcons->setSavedValue(SDI_GLOW, glow);
             else gameManager->setPlayerGlow(glow);
             return glow;
         case ICON_RANDOMIZER_API_CUBE:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_CUBE, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_CUBE, num);
                 else gameManager->setPlayerFrame(num);
                 MoreIcons::setIcon("", IconType::Cube, useDual);
                 return num;
@@ -270,7 +261,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_SHIP:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_SHIP, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_SHIP, num);
                 else gameManager->setPlayerShip(num);
                 MoreIcons::setIcon("", IconType::Ship, useDual);
                 return num;
@@ -279,7 +270,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_BALL:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_BALL, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_BALL, num);
                 else gameManager->setPlayerBall(num);
                 MoreIcons::setIcon("", IconType::Ball, useDual);
                 return num;
@@ -288,7 +279,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_UFO:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_UFO, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_UFO, num);
                 else gameManager->setPlayerBird(num);
                 MoreIcons::setIcon("", IconType::Ufo, useDual);
                 return num;
@@ -297,7 +288,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_WAVE:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_WAVE, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_WAVE, num);
                 else gameManager->setPlayerDart(num);
                 MoreIcons::setIcon("", IconType::Wave, useDual);
                 return num;
@@ -306,7 +297,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_ROBOT:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_ROBOT, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_ROBOT, num);
                 else gameManager->setPlayerRobot(num);
                 MoreIcons::setIcon("", IconType::Robot, useDual);
                 return num;
@@ -315,7 +306,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_SPIDER:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_SPIDER, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_SPIDER, num);
                 else gameManager->setPlayerSpider(num);
                 MoreIcons::setIcon("", IconType::Spider, useDual);
                 return num;
@@ -324,7 +315,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_SWING:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_SWING, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_SWING, num);
                 else gameManager->setPlayerSwing(num);
                 MoreIcons::setIcon("", IconType::Swing, useDual);
                 return num;
@@ -333,7 +324,7 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             return 0;
         case ICON_RANDOMIZER_API_JETPACK:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_JETPACK, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_JETPACK, num);
                 else gameManager->setPlayerJetpack(num);
                 MoreIcons::setIcon("", IconType::Jetpack, useDual);
                 return num;
@@ -341,12 +332,12 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             MoreIcons::setIcon(loadedIcons[num], IconType::Jetpack, useDual);
             return 0;
         case ICON_RANDOMIZER_API_DEATH_EFFECT:
-            if (useDual) setModValue(separateDualIcons, SDI_DEATH, num);
+            if (useDual) separateDualIcons->setSavedValue(SDI_DEATH, num);
             else gameManager->setPlayerDeathEffect(num);
             return num;
         case ICON_RANDOMIZER_API_TRAIL:
             if (idx < vec.size()) {
-                if (useDual) setModValue(separateDualIcons, SDI_TRAIL, num);
+                if (useDual) separateDualIcons->setSavedValue(SDI_TRAIL, num);
                 else gameManager->setPlayerStreak(num);
                 MoreIcons::setIcon("", IconType::Special, useDual);
                 return num;
@@ -354,9 +345,13 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
             MoreIcons::setIcon(loadedIcons[num], IconType::Special, useDual);
             return 0;
         case ICON_RANDOMIZER_API_SHIP_FIRE:
-            if (useDual) setModValue(separateDualIcons, SDI_SHIP_FIRE, num);
+            if (useDual) separateDualIcons->setSavedValue(SDI_SHIP_FIRE, num);
             else gameManager->setPlayerShipStreak(num);
             return num;
+        case ICON_RANDOMIZER_API_EXPLODE:
+            if (useDual) separateDualIcons->setSavedValue(SDI_EXPLODE, explode);
+            else gameManager->setGameVariable("0153", explode);
+            return explode;
         default:
             return -1;
     }
@@ -381,6 +376,7 @@ void IconRandomizer::randomizeAll(RandomizeAllType type, bool dual) {
             randomize(ICON_RANDOMIZER_API_TRAIL, dual);
             randomize(ICON_RANDOMIZER_API_SHIP_FIRE, dual);
             randomize(ICON_RANDOMIZER_API_ANIMATION, dual);
+            randomize(ICON_RANDOMIZER_API_EXPLODE, dual);
             break;
         case ICON_RANDOMIZER_API_ALL_COLORS:
             randomize(ICON_RANDOMIZER_API_COLOR_1, dual);
@@ -394,44 +390,46 @@ void IconRandomizer::randomizeAll(RandomizeAllType type, bool dual) {
 int IconRandomizer::active(RandomizeType type, bool dual) {
     // Get the game manager, then get the Separate Dual Icons mod.
     // If the mod is enabled and the dual parameter is true, use separate dual icons.
-    auto gameManager = GameManager::sharedState();
+    auto gameManager = GameManager::get();
     auto separateDualIcons = Loader::get()->getLoadedMod(SDI_ID);
     auto useDual = separateDualIcons && dual;
 
     // Get the value from the Separate Dual Icons mod if it is enabled and the dual parameter is true, otherwise get the value from the game manager.
     switch (type) {
         case ICON_RANDOMIZER_API_COLOR_1:
-            return useDual ? getModValue(separateDualIcons, SDI_COLOR_1, 0) : gameManager->getPlayerColor();
+            return useDual ? separateDualIcons->getSavedValue(SDI_COLOR_1, 0) : gameManager->getPlayerColor();
         case ICON_RANDOMIZER_API_COLOR_2:
-            return useDual ? getModValue(separateDualIcons, SDI_COLOR_2, 0) : gameManager->getPlayerColor2();
+            return useDual ? separateDualIcons->getSavedValue(SDI_COLOR_2, 0) : gameManager->getPlayerColor2();
         case ICON_RANDOMIZER_API_GLOW_COLOR:
-            return useDual ? getModValue(separateDualIcons, SDI_GLOW_COLOR, 0) : gameManager->getPlayerGlowColor();
+            return useDual ? separateDualIcons->getSavedValue(SDI_GLOW_COLOR, 0) : gameManager->getPlayerGlowColor();
         case ICON_RANDOMIZER_API_GLOW:
-            return useDual ? getModValue(separateDualIcons, SDI_GLOW, false) : gameManager->getPlayerGlow();
+            return useDual ? separateDualIcons->getSavedValue(SDI_GLOW, false) : gameManager->getPlayerGlow();
         case ICON_RANDOMIZER_API_CUBE:
-            return useDual ? getModValue(separateDualIcons, SDI_CUBE, 1) : gameManager->getPlayerFrame();
+            return useDual ? separateDualIcons->getSavedValue(SDI_CUBE, 1) : gameManager->getPlayerFrame();
         case ICON_RANDOMIZER_API_SHIP:
-            return useDual ? getModValue(separateDualIcons, SDI_SHIP, 1) : gameManager->getPlayerShip();
+            return useDual ? separateDualIcons->getSavedValue(SDI_SHIP, 1) : gameManager->getPlayerShip();
         case ICON_RANDOMIZER_API_BALL:
-            return useDual ? getModValue(separateDualIcons, SDI_BALL, 1) : gameManager->getPlayerBall();
+            return useDual ? separateDualIcons->getSavedValue(SDI_BALL, 1) : gameManager->getPlayerBall();
         case ICON_RANDOMIZER_API_UFO:
-            return useDual ? getModValue(separateDualIcons, SDI_UFO, 1) : gameManager->getPlayerBird();
+            return useDual ? separateDualIcons->getSavedValue(SDI_UFO, 1) : gameManager->getPlayerBird();
         case ICON_RANDOMIZER_API_WAVE:
-            return useDual ? getModValue(separateDualIcons, SDI_WAVE, 1) : gameManager->getPlayerDart();
+            return useDual ? separateDualIcons->getSavedValue(SDI_WAVE, 1) : gameManager->getPlayerDart();
         case ICON_RANDOMIZER_API_ROBOT:
-            return useDual ? getModValue(separateDualIcons, SDI_ROBOT, 1) : gameManager->getPlayerRobot();
+            return useDual ? separateDualIcons->getSavedValue(SDI_ROBOT, 1) : gameManager->getPlayerRobot();
         case ICON_RANDOMIZER_API_SPIDER:
-            return useDual ? getModValue(separateDualIcons, SDI_SPIDER, 1) : gameManager->getPlayerSpider();
+            return useDual ? separateDualIcons->getSavedValue(SDI_SPIDER, 1) : gameManager->getPlayerSpider();
         case ICON_RANDOMIZER_API_SWING:
-            return useDual ? getModValue(separateDualIcons, SDI_SWING, 1) : gameManager->getPlayerSwing();
+            return useDual ? separateDualIcons->getSavedValue(SDI_SWING, 1) : gameManager->getPlayerSwing();
         case ICON_RANDOMIZER_API_JETPACK:
-            return useDual ? getModValue(separateDualIcons, SDI_JETPACK, 1) : gameManager->getPlayerJetpack();
+            return useDual ? separateDualIcons->getSavedValue(SDI_JETPACK, 1) : gameManager->getPlayerJetpack();
         case ICON_RANDOMIZER_API_DEATH_EFFECT:
-            return useDual ? getModValue(separateDualIcons, SDI_DEATH, 1) : gameManager->getPlayerDeathEffect();
+            return useDual ? separateDualIcons->getSavedValue(SDI_DEATH, 1) : gameManager->getPlayerDeathEffect();
         case ICON_RANDOMIZER_API_TRAIL:
-            return useDual ? getModValue(separateDualIcons, SDI_TRAIL, 1) : gameManager->getPlayerStreak();
+            return useDual ? separateDualIcons->getSavedValue(SDI_TRAIL, 1) : gameManager->getPlayerStreak();
         case ICON_RANDOMIZER_API_SHIP_FIRE:
-            return useDual ? getModValue(separateDualIcons, SDI_SHIP_FIRE, 1) : gameManager->getPlayerShipFire();
+            return useDual ? separateDualIcons->getSavedValue(SDI_SHIP_FIRE, 1) : gameManager->getPlayerShipFire();
+        case ICON_RANDOMIZER_API_EXPLODE:
+            return useDual ? separateDualIcons->getSavedValue(SDI_EXPLODE, 0) : gameManager->getGameVariable("0153");
         default:
             return -1;
     }